Listo

+++ app/src/main/java/com/example/ejerciciotiempo/Modelo/City.java
@@ -1,17 +1,17 @@
 package com.example.ejerciciotiempo.Modelo;
 
 public class City {
-    private float temp;
-    private float feels_like;
-    private float temp_min;
-    private float temp_max;
-    private float pressure;
-    private float humidity;
+    private double temp;
+    private double feels_like;
+    private double temp_min;
+    private double temp_max;
+    private double pressure;
+    private double humidity;
 
     public City() { }
 
-    public City(float temp, float feels_like, float temp_min,
-                     float temp_max, float pressure, float humidity) {
+    public City(double temp, double feels_like, double temp_min,
+                     double temp_max, double pressure, double humidity) {
         this.temp = temp;
         this.feels_like = feels_like;
         this.temp_min = temp_min;
@@ -20,51 +20,51 @@ public class City {
         this.humidity = humidity;
     }
 
-    public float getTemp() {
+    public double getTemp() {
         return temp;
     }
 
-    public void setTemp(float temp) {
+    public void setTemp(double temp) {
         this.temp = temp;
     }
 
-    public float getFeels_like() {
+    public double getFeels_like() {
         return feels_like;
     }
 
-    public void setFeels_like(float feels_like) {
+    public void setFeels_like(double feels_like) {
         this.feels_like = feels_like;
     }
 
-    public float getTemp_min() {
+    public double getTemp_min() {
         return temp_min;
     }
 
-    public void setTemp_min(float temp_min) {
+    public void setTemp_min(double temp_min) {
         this.temp_min = temp_min;
     }
 
-    public float getTemp_max() {
+    public double getTemp_max() {
         return temp_max;
     }
 
-    public void setTemp_max(float temp_max) {
+    public void setTemp_max(double temp_max) {
         this.temp_max = temp_max;
     }
 
-    public float getPressure() {
+    public double getPressure() {
         return pressure;
     }
 
-    public void setPressure(float pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public float getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(float humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
 }
